Add header, update styles to roughly match main site
It's very disorienting trying to navigate a site where half of the links
on the navbar take you to a different site with a completely different
navbar.

This fix copies the navbar from the main site (removing the ikiwiki-
generated breadcrumbs) and copies some of the styles. It's not ideal
(our website needs some work) but at the very least it's possible to
navigate without needing to use the back button constantly.

The actual HTML and CSS changes are quite nasty but seem to work
reasonably well as a temporary fix until we eventually redo the entire
site.