webapps/nodejs: remove socket on start, umask to 0
nodejs throws EADDRINUSE without doing anything else if a socket already
exists, even if nothing is bound to it (other servers tend to behave
sanely).

It also gives overly restrictive permissions, preventing the webserver
from writing to it. umask 0 is the best solution I could find to that,
but it's a little dangerous. Ideally groups would umask down again after
the server starts.

Open to other solutions in the meantime...